Updated example, test config and ts imports
