+++ example/package.json
@@ -15,12 +15,12 @@
   "dependencies": {
     "@types/google-protobuf": "^3.2.5",
     "google-protobuf": "^3.2.0",
-    "grpc-web-client": "0.1.3"
+    "grpc-web-client": "0.2.0"
   },
   "devDependencies": {
     "concurrently": "^3.4.0",
     "ts-loader": "^2.0.1",
-    "ts-protoc-gen": "^0.1.1",
+    "ts-protoc-gen": "^0.2.1",
     "typescript": "^2.2.1",
     "webpack": "^2.2.1",
     "webpack-dev-server": "^2.4.2"
