+++ example/ts/src/index.ts
@@ -1,5 +1,5 @@
 import {grpc, BrowserHeaders} from "grpc-web-client";
-import {BookService} from "./services";
+import {BookService} from "../_proto/examplecom/library/book_service_pb_service";
 import {QueryBooksRequest, Book, GetBookRequest} from "../_proto/examplecom/library/book_service_pb";
 
 declare const USE_TLS: boolean;
@@ -16,9 +16,6 @@ function getBook() {
     },
     onMessage: function (message: Book) {
       console.log("getBook.onMessage", message.toObject());
-    },
-    onError: function (err: Error) {
-      console.error("getBook.onError", err);
     },
     onComplete: function (code: grpc.Code, msg: string | undefined, trailers: BrowserHeaders) {
       console.log("getBook.onComplete", code, msg, trailers);
@@ -41,9 +38,6 @@ function queryBooks() {
     },
     onMessage: function(message: Book) {
       console.log("queryBooks.onMessage", message.toObject());
-    },
-    onError: function(err: Error) {
-      console.error("queryBooks.onError", err);
     },
     onComplete: function(code: grpc.Code, msg: string | undefined, trailers: BrowserHeaders) {
       console.log("queryBooks.onComplete", code, msg, trailers);
