+++ test/package.json
@@ -9,7 +9,7 @@
     "build:dev": "cd ts && rm -rf build && webpack --watch",
     "build": "npm run build:proto && npm run build:testserver && npm run build:ts",
     "lint": "tslint -c ./ts/tslint.json ./ts/src/**/*.ts",
-    "test": "npm run build && npm run test:http && npm run test:https",
+    "test": "npm run build && npm run test:https && npm run test:http",
     "test:http": "./run-karma.sh ./http.karma.conf.js --single-run",
     "test:https": "./run-karma.sh ./https.karma.conf.js --single-run",
     "test:dev:http": "./run-karma.sh ./http.karma.conf.js",
