+++ ts/README.md
@@ -21,9 +21,6 @@ grpc.invoke(TestService.Ping, {
   onMessage: function(message: PingResponse) {
 
   },
-  onError: function(err: Error) {
-
-  },
   onComplete: function(code: grpc.Code, msg: string | undefined, trailers: BrowserHeaders) {
 
   }
